requirements.txt
Django==4.0.2
django-mysql==4.5.0
mysqlclient==2.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
pyjwt==2.3.0